Cell Detection Complete
Cells now are aware of all adjacent cells in all NSEW directions.

A list of cells to avoid can now also be passed in to the Maze object
which then filters down to the individual tiles to be ignored when
finding suitable adjacent cells. This appears to be almost working
however cells dont seem to be ignoring cells beneath them... not sure
whats causing this.