New Display and Working on Algorithms.. finally!
The maze now has some sort of visual element to it, although it will
eventually become a <canvas> GUI. I have added the ability for cells to
know if they are avoided or not, which let me refactor the code that was
used to work out if a cell should be avoided or not.

On top of that I have started to implement the first algorithm for
actually making the mazes, which at the moment stands at: ~working~. It
creates some sort of a maze looking.. thing, but using only a handful of
cells. I believe this is a problem with the algorithm picking cells its
already used and burning itself out too fast. Need to look in to it.

Lastly, in the process of making all these new additions and changes, I
have created a handful of new functions for the classes that were
required. Including two static functions in the Position class.