Now we're generating!
The code for generating the maze is now complete. Its looking pretty
good to be honest, even if the graphics are little more than a couple
minutes in paint.
Next step is to start work on saving the completed maze to json for
saving/using on an actual webpage.

Changes:
Completely re-wrote the algorithm so that it actually made sense and
worked.
Added another function to the Cell class that is used to return the
positions only for an array of cells (extremely handy for debugging).