JSON Encode added
Once the maze is generated it can now be exported via a JSON string.
This would then be used in javascript to create the final product, or
any software really. I decided not to work on loaded of already
generated mazes since that isn't really the point of the scripts. Only
generating them. Even if you could import a JSON string and have the
maze back in the correct objects and such, there isn't really anything
you could do with it.

Time to move on to creating the visual element of the project and put
some thought in to what exactly I want to do with it.