Learning HTML5 Canvas
The task at the moment is making something that actually uses the mazes
that I can now generate.

The first task is being about to convert the JSON to a viewable maze in
the webpage and have some interaction with it. Then move on to being
able to step through the maze with perhaps some sort of step tracker.
Once that is working I can begin making it actually interesting to play
around with.