Bug fix and Canvas update
Working on the new display meant creating smaller mazes that highlighted
a bug in the generation code. Although avoided cells were rightfully
disregarded in the generation, it didn't stop the very first cell that
gets picked being on the avoided list that lead to some broken designs.
Bug now fixed.

The canvas work has grown in size with the javascript seeing new
functions, redesigns of how the cells are drawn and in the importing of
the maze json (although not finished). Without the internet and
therefore without jQuery this has been a lot harder. For somethings Im
just going to have to wait.

The maze now is drawn, centered around a randomly selected cell (doesn't
check if the cell is avoided yet) and there are basic up/down/left/right
controls, albeit backwards due to me moving the background rather than
the player.

As ever, always things to fix.